requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
PyJWT==2.8.0
bcrypt==4.0.1
openai>=1.50.0
python-dotenv==1.0.0
gunicorn==21.2.0
Flask-Bcrypt==1.0.1
